kylin-order:初步实现GET order/foods接口

+++ kylin-order/src/main/java/com/store59/kylin/order/data/mapper/OrderfoodMapper.java
@@ -17,5 +17,5 @@ public interface OrderfoodMapper {
 
 	int updateByPrimaryKey(Orderfood record);
 
-    List<Orderfood> selectByOrderId(Integer orderId);
+    List<Orderfood> selectByOrderId(Long orderId);
 }
